Fix wrongfully called destructor issue in mCRL2 greybox
The context structure in the mCRL2 greybox is being malloc'ed instead
of new'ed, which means the contructors for the objects in the
structure weren't being called. However, the destructor was called at
some point resulting in a segfault in the mpi version of the tool. Fix
this by replacing the objects in the structure by pointers and using
new where needed. Calling new on the structure directly is not
possible as the plain constructors for the objects in the structure
are protected.