Ensure the last step in a trace to an action is present
* Both states are passed to the find trace function instead
  of only the destination state.
* This also removes the destruction of the levels that took
  place in the action finding code.

To explain: the code with which we skip levels would also
skip the action steps in which we are interested in case
the step is effectively a self-loop. The current code makes
sure both states will occur at the end of the trace being
constructed by passing both states to the find_trace
function and then always adding them to the trace.

The code is a bit more generic and allows an arbitrary
sequence of states being passed.

(Closes #429)

Signed-off-by: Michael Weber <michaelw@cs.utwente.nl>