Add option -v to --mcrl2 options, which increases verbosity of mcrl2
Adds options --help and --usage, and short option -r as
alternative to --rewriter.

Also clean up a bit of const mess along the way.

NOTE: since we parse all options anyway, the option handling
could be moved to the outer level and we get rid of --mcrl2
entirely.  Not done here for backwards compatibility.