Fix mCRL2 interface with pins_open
The mCRL2 pins_open interface was broken before the latest
work on mCRL2 linking (stdc++ was never linked against).
Linking against stdc++ is now explicit; this is not a
100% correct fix, but seems to work for gcc.

Pre-processing strings from CADP .c files was also broken.
This now works in case paths do not contain spaces. Note
that the script is broken irrespectively of this in case
paths with spaces in them are used.