New better LMAP!
lmap.h functions like the compact hash table of
Geldenhuys & Valmari and uses a very simple
integrated allocator to hand out blocks.
The blocksize B is configurable, with B=2 it
degrades to a linked list, B=3 is memory optimal
in the compact hash table setting and B=8
is optimal wrt performance (block on the cache line)
Real memory requirements depend on the size of the
different sets associated with the states.

Improvements over lattice-map
* saves memory per lattice (1 word io 2 words)
* there was a bug in lattice-map deletion
* cleaner code
* locks can be integrated
* smaller memory footprint (can be optimized with prefetch)

TODO:
* add prefetch