Extend SPIN LTL semantics to deadlocks. Renamed old semantics to "ltsmin"
SPIN and DiVinE have special LTL semantics for deadlocks
in the LTS. Here, the buchi is allowed to keep running
unchecked. The patch extends the LTL layer with
this behavior and also updates the dependency matrix.
The old behavior is morei nteresting for LTSmin's LTL
algorithms and is therefore preserved under the name
"ltsmin" (see --ltl-semantics)