Support progress transitions and states
Uses static progress information on groups, via the transition label "statement", which encodes progress if it contains the text "<progress>"

Falls back on progress states when this info is not available. In this case the state label "progress" is used and all progress states are fully expanded in POR. Additionally the visibility of the progress label is exported and used, so that POR is guided to avoid progress.

PINS visibility matrix and POR visibility in tools