Relax POR for safety properties
- Disable visibility proviso (while still
  recording visible transitions / labels)
- Implement safety proviso, by selecting all
  visible transitions if ignoring is detected.
  The POR search makes sure that all ignored
  transitions are selected which could lead to
  visible transitions

The best way to implement the safety proviso
would be to precisely compute ignored transitions
in bottom SCCs of the reduced state space and
include one in the POR search (See Valmari).
This however would couple the POR layer to tightly
to the algorithms.
The current approach seems to exploit the
available infra structure best and is still
better than Peled's proviso.