Implement minimal cost reachability in pins2lts-dist.
This patch add an option that can select an edge label to
define the cost of edges. This cost must be non-negative.
The tool will explore successively higher cost levels.
Each of these levels is store in a doubly linked list.
First, because this means O(1) remove and insert when
a lower cost path is found. Second, because the order
within a cost level remains close to BFS and therefore
should not be excessively long.