Change PINS API for GBgroupsOfEdge.
Previously it was necessary to do a malloc for every call to
GBgroupsOfEdge(). This is unnecessarily expensive. Now the API is changed
in such a way that the groups parameter is a const int**. This means
that the language front-end has to supply a pointer to the integer
array indicating all groups that can produce the edge, previously the
back-end would supply this pointer. Also, the back-end is unable to free
the pointer. Meaning the best way to implement this method now is by
doing a malloc once, store the obtained pointer in the model_t user context
and write the necessary groups every time a GBgroupsOfEdge call is done.
This means said pointer should point to a block that is at least n
integers long, where n is the number of state slots.