Add code that wraps around native qsort_r functions.
This is now possible because _DARWIN_C_SOURCE is now defined.

To elaborate on this, defining _XOPEN_SOURCE hides qsort_r, but
defining _DARWIN_C_SOURCE enables qsort_r again. The wrapper code
is necessary to have a common function signature for qsort_r; the
order of arguments is different between BSD and GNU.

Addresses #109.