Pass proper compiler define for jittyc rewriter to mCRL2.
When compiling with mCRL2 support, and when support for the jittyc rewriter is not explicitly disabled, pass the compiler define MCRL2_JITTYC_AVAILABLE as a preprocessor flag.
This is required to enable the support for the jittyc rewriter in mCRL2.

It seems that currently there is no way to determine directly from an mCRL2 installation whether it supports the jittyc rewriter. I will discuss on the mCRL2 side whether this can be improved. However, at least for the 202106.0 release of mCRL2 and the current development version, this change is required. I have not tested with older releases.