remove all whitespace from predicate keys used in lookup
In HOA mode AP without whitespace were compared to AP in le_list with
(weird) whitespaces in them. This patch forces all these strings to be
ws free.

Also revise the strategy for parse of AP in a HOA; they should now
simply be parsable predicates in LTSmin syntax, instead of names of
state labels. We now build a trivial formula with these AP to mimic the
"normal" LTL workflow more accurately. This helps with all the
registration steps.